+++ Estimation/Merge/UMIs/MergeUMIsStrategySimple.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <Estimation/Cell.h>
-#include <Estimation/CellsDataContainer.h>
+#include "MergeUMIsStrategyAbstract.h"
 
 #include <unordered_set>
 #include <unordered_map>
@@ -18,7 +17,7 @@ namespace Estimation
 	{
 		namespace UMIs
 		{
-			class MergeUMIsStrategySimple
+			class MergeUMIsStrategySimple : public MergeUMIsStrategyAbstract
 			{
 				friend struct TestEstimator::testFillWrongUmis;
 				friend struct TestEstimator::testRemoveSimilarWrongUmis;
@@ -26,6 +25,7 @@ namespace Estimation
 			private:
 				using s_hash_t = std::unordered_set<std::string>;
 				using s_vec_t = std::vector<std::string>;
+
 			private:
 				const unsigned _max_merge_distance;
 				static const std::string nucleotides;
@@ -36,11 +36,11 @@ namespace Estimation
 				void remove_similar_wrong_umis(s_vec_t &wrong_umis) const;
 				CellsDataContainer::s_s_hash_t fill_wrong_umis(s_vec_t &wrong_umis) const;
 
-			public:
-				MergeUMIsStrategySimple(unsigned max_merge_distance);
-				void merge(CellsDataContainer &container) const;
-
 				bool is_umi_real(const std::string &umi) const;
+
+			public:
+				explicit MergeUMIsStrategySimple(unsigned max_merge_distance);
+				void merge(CellsDataContainer &container) const override;
 			};
 		}
 	}